Co2-Sensor auslesen.

+++ src/AirMate/Sensors.h
@@ -5,7 +5,7 @@ extern bool g_autoEnabled;
 
 extern float g_temperature;
 extern float g_humidity;
-extern float g_co2ppm;
+extern int32_t g_co2ppm;
 
 void sensors_setup();
 void sensors_update();
